Extract RowRow and UserRow components
